Add message-passing for flash interaction
The script context isn't allowed to call plugin methods on chrome any
more, so need to run code in the page context to interact with Flash,
and use message passing between that and the main script.

New util wrappers for everything, with async callbacks instead of return
values, because of course we need that now.